fix: update CI configuration and improve coverage report generation refactor: optimize embedding model implementation and improve code clarity refactor: streamline agent command execution and remove unused imports refactor: enhance benchmark command structure for better readability fix: correct find command argument handling for elapsed time refactor: simplify process command logic and improve data handling refactor: update UI components to remove dead code and improve clarity

+++ xore-cli/src/commands/find.rs
@@ -239,7 +239,7 @@ pub fn execute(args: FindArgs) -> Result<()> {
             search_type,
             &args.path,
             sorted_files.len(),
-            stats.elapsed_ms as u64,
+            stats.elapsed_ms,
             args.file_type.clone(),
         ) {
             Ok(_) => {
